--- src/Features/Authentication/Login/temp/SeniorUIUX.Anastasia.module.scss
@@ -1,7 +1,0 @@
-.NastyaSuper {
-    display: flex;
-}
-
-.Shopping {
-    position: relative;
-}
