+++ src/Shared/Api/baseQuery.ts
@@ -1,3 +1,6 @@
+import { config } from '../Config/config'
+import { getAccessToken } from '../Lib/Helpers/getAccessToken'
+
 import {
     BaseQueryFn,
     FetchArgs,
@@ -5,9 +8,6 @@ import {
     FetchBaseQueryMeta,
     fetchBaseQuery,
 } from '@reduxjs/toolkit/dist/query/react'
-import { config } from './config'
-
-import { getAccessToken } from '../Lib/Helpers/getAccessToken'
 
 export const baseQuery: BaseQueryFn<
     string | FetchArgs,
