+++ src/Shared/Api/index.ts
@@ -1,5 +1,5 @@
 export { baseApi } from './baseApi'
 export { SESSION_TAG, USER_TAG, CATEGORIES_TAG, TASK_TAG } from './tags'
-export { config } from './config'
 export { invalidateAccessToken } from './invalidateTokenEvent'
+
 export * from './isFetchBaseQueryError'
